use Decimal.decimal for amounts

+++ examples/create_non_std_tx.py
@@ -10,6 +10,7 @@
 # propagated, or distributed except according to the terms contained in the
 # LICENSE file.
 
+from decimal import Decimal
 from bitcoinutils.setup import setup
 from bitcoinutils.transactions import Transaction, TxInput, TxOutput
 from bitcoinutils.keys import P2pkhAddress, PrivateKey
@@ -33,7 +34,7 @@ def main():
     txin = TxInput('e2d08a63a540000222d6a92440436375d8b1bc89a2638dc5366833804287c83f', 1)
 
     # locking script expects 2 numbers that when added equal 5 (silly example)
-    txout = TxOutput(int(0.9 * COIN), Script(['OP_ADD', 'OP_5', 'OP_EQUAL']) )
+    txout = TxOutput(Decimal(0.9 * COIN), Script(['OP_ADD', 'OP_5', 'OP_EQUAL']) )
 
     # create another output to get the change - remaining 0.01 is tx fees
     # note that this time we used to_script_pub_key() to create the P2PKH
